--- src/menu-items/dashboard.js
@@ -1,28 +1,0 @@
-// assets
-import { IconDashboard, IconDeviceAnalytics } from '@tabler/icons';
-
-// constant
-const icons = {
-    IconDashboard,
-    IconDeviceAnalytics
-};
-
-// ===========================|| DASHBOARD MENU ITEMS ||=========================== //
-
-const dashboard = {
-    id: 'dashboard',
-    title: 'Dashboard',
-    type: 'group',
-    children: [
-        {
-            id: 'default',
-            title: 'Dashboard',
-            type: 'item',
-            url: '/dashboard/default',
-            icon: icons.IconDashboard,
-            breadcrumbs: false
-        }
-    ]
-};
-
-export default dashboard;
