+++ src/menu-items/index.js
@@ -1,4 +1,4 @@
-import dashboard from './dashboard';
+import clinical from './clinical';
 // import pages from './pages';
 import utilities from './utilities';
 import other from './other';
@@ -6,7 +6,7 @@ import other from './other';
 // ===========================|| MENU ITEMS ||=========================== //
 
 const menuItems = {
-    items: [dashboard, /* pages, */ utilities, other]
+    items: [clinical, /* pages, */ utilities, other]
 };
 
 export default menuItems;
